refactor(domain): 로또와 로또구입정보에 대한 코드를 domain 패키지로 분리
