+++ src/main/java/lotto/domain/LottoPurchaseInfo.java
@@ -1,0 +1,11 @@
+package lotto.domain;
+
+public class LottoPurchaseInfo {
+    int purchaseAmount;
+    int purchaseLottoNum;
+
+    public LottoPurchaseInfo(int purchaseAmount, int purchaseLottoNum) {
+        this.purchaseAmount = purchaseAmount;
+        this.purchaseLottoNum = purchaseLottoNum;
+    }
+}
